refactor: 채팅 파일 1차 리팩토링 (#47)
