+++ src/features/chat/components/index.ts
@@ -1,23 +1,23 @@
-export { ChatPageContent } from "@/features/chat/components/ChatPageContent";
-export type { ChatPageContentProps } from "@/features/chat/components/ChatPageContent";
 export { ChatDetailContent } from "@/features/chat/components/ChatDetailContent";
 export type { ChatDetailContentProps } from "@/features/chat/components/ChatDetailContent";
-export { ChatDateBadge } from "@/features/chat/components/ChatDateBadge";
-export type { ChatDateBadgeProps } from "@/features/chat/components/ChatDateBadge";
-export { ChatInputMenu } from "@/features/chat/components/ChatInputMenu";
-export type { ChatInputMenuProps } from "@/features/chat/components/ChatInputMenu";
+export { ChatDateBadge } from "@/features/chat/components/chat-detail/ChatDateBadge";
+export type { ChatDateBadgeProps } from "@/features/chat/components/chat-detail/ChatDateBadge";
+export { ChatMessageList } from "@/features/chat/components/chat-detail/ChatMessageList";
+export type { ChatMessageListProps } from "@/features/chat/components/chat-detail/ChatMessageList";
+export { ChatInputMenu } from "@/features/chat/components/chat-detail/ChatInputMenu";
+export type { ChatInputMenuProps } from "@/features/chat/components/chat-detail/ChatInputMenu";
 export type { ChatInputMenuAction } from "@/features/chat/types";
-export { ChatRoommateInviteSheet } from "@/features/chat/components/ChatRoommateInviteSheet";
-export type { ChatRoommateInviteSheetProps } from "@/features/chat/components/ChatRoommateInviteSheet";
+export { ChatRoommateInviteSheet } from "@/features/chat/components/chat-detail/ChatRoommateInviteSheet";
+export type { ChatRoommateInviteSheetProps } from "@/features/chat/components/chat-detail/ChatRoommateInviteSheet";
 export { ChatRoommateInviteMessage } from "@/features/chat/components/ChatRoommateInviteMessage";
 export type { ChatRoommateInviteMessageProps } from "@/features/chat/components/ChatRoommateInviteMessage";
-export { ChatRoommateConfirmedContent } from "@/features/chat/components/ChatRoommateConfirmedContent";
-export type { ChatRoommateConfirmedContentProps } from "@/features/chat/components/ChatRoommateConfirmedContent";
+export { ChatRoommateConfirmedContent } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
+export type { ChatRoommateConfirmedContentProps } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
 export { ChatRoommateRequestMessage } from "@/features/chat/components/ChatRoommateRequestMessage";
 export type { ChatRoommateRequestMessageProps } from "@/features/chat/components/ChatRoommateRequestMessage";
-export { ChatMatchCard } from "@/features/chat/components/ChatMatchCard";
-export type { ChatMatchCardProps } from "@/features/chat/components/ChatMatchCard";
-export { ChatListItem } from "@/features/chat/components/ChatListItem";
-export type { ChatListItemProps } from "@/features/chat/components/ChatListItem";
-export { MessageCount } from "@/features/chat/components/MessageCount";
-export type { MessageCountProps } from "@/features/chat/components/MessageCount";
+export { ChatMatchCard } from "@/features/chat/components/chat-detail/ChatMatchCard";
+export type { ChatMatchCardProps } from "@/features/chat/components/chat-detail/ChatMatchCard";
+export { ChatListItem } from "@/features/chat/components/chat-list/ChatListItem";
+export type { ChatListItemProps } from "@/features/chat/components/chat-list/ChatListItem";
+export { MessageCount } from "@/features/chat/components/chat-list/MessageCount";
+export type { MessageCountProps } from "@/features/chat/components/chat-list/MessageCount";
