+++ src/features/chat/index.ts
@@ -1,9 +1,9 @@
-export { ChatPageContent } from "@/features/chat/components/ChatPageContent";
-export type { ChatPageContentProps } from "@/features/chat/components/ChatPageContent";
 export { ChatDetailContent } from "@/features/chat/components/ChatDetailContent";
 export type { ChatDetailContentProps } from "@/features/chat/components/ChatDetailContent";
-export { ChatRoommateConfirmedContent } from "@/features/chat/components/ChatRoommateConfirmedContent";
-export type { ChatRoommateConfirmedContentProps } from "@/features/chat/components/ChatRoommateConfirmedContent";
+export { ChatListItem } from "@/features/chat/components/chat-list/ChatListItem";
+export type { ChatListItemProps } from "@/features/chat/components/chat-list/ChatListItem";
+export { ChatRoommateConfirmedContent } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
+export type { ChatRoommateConfirmedContentProps } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
 export {
   useChatComposer,
   useChatMessages,
@@ -11,12 +11,13 @@ export {
   useCreateChatRoom,
   useIssueChatWsToken,
 } from "@/features/chat/hooks";
-export { CHAT_DETAILS } from "@/features/chat/mocks";
+export { CHAT_DETAILS, CHAT_HELPER_TEXT, CHAT_PREVIEWS, CHAT_TABS } from "@/features/chat/mocks";
 export type {
   ChatConnectionStatus,
   ChatErrorMessage,
   ChatReceivedMessage,
   ChatSendMessagePayload,
+  ChatTab,
   ChatUserProfile,
 } from "@/features/chat/types";
 export {
@@ -24,3 +25,4 @@ export {
   isChatErrorMessage,
   isChatReceivedMessage,
 } from "@/features/chat/utils/webSocket";
+export { mapHistoryMessagesToChatMessages } from "@/features/chat/utils/chatHistoryMessages";
