+++ src/features/chat/sections/chat-detail/index.ts
@@ -1,0 +1,2 @@
+export { ChatMessageListSection } from "@/features/chat/sections/chat-detail/ChatMessageListSection";
+export type { ChatMessageListSectionProps } from "@/features/chat/sections/chat-detail/ChatMessageListSection";
