+++ src/features/chat/sections/index.ts
@@ -1,0 +1,2 @@
+export { ChatMessageListSection } from "@/features/chat/sections/chat-detail";
+export type { ChatMessageListSectionProps } from "@/features/chat/sections/chat-detail";
