--- src/pages/ChatPage.tsx
@@ -1,5 +1,0 @@
-import { ChatPageContent } from "@/features/chat";
-
-export default function ChatPage() {
-  return <ChatPageContent />;
-}
