+++ src/pages/chat/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useChatDetailPage } from "@/pages/chat/hooks/useChatDetailPage";
